--- frontend/src/Components/MapContextForm/MapContextForm.css
@@ -1,12 +1,0 @@
-.map-context-layout {
-  width: calc(
-    100% + 48px
-  ); /* to override parent's padding defined in site-layout class */
-  height: calc(
-    100% + 48px
-  ); /* to override parent's padding defined in site-layout class */
-  position: relative;
-  margin-left: -24px; /* to override parent's padding defined in site-layout class */
-  margin-top: -24px; /* to override parent's padding defined in site-layout class */
-  max-height: calc(100vh);
-}
