--- frontend/src/Components/TheMap/LayerManager/LayerManager.css
@@ -1,21 +1,0 @@
-.layer-manager-toggle {
-  top:50%;
-  position: absolute;
-  z-index: 1000;
-  height: 60px;
-  width: 30px;
-  padding: 0;
-  border-radius: 0 5px 5px 0;
-  background-color: rgba(0, 0, 0, 0.85);
-  border: unset;
-  /* same transition timing as drawer */
-  transition: left 0.2s cubic-bezier(0.23, 1, 0.32, 1);
-}
-
-.layer-manager-toggle:focus{
-  background-color: rgba(0, 0, 0, 0.85);
-}
-
-.layer-manager-toggle:hover{
-  background-color: rgba(0, 0, 0, 0.85);
-}
