--- frontend/src/Components/TheMap/TheMap.css
@@ -1,11 +1,0 @@
-.the-map-container {
-  display: flex;
-  height: 100%;
-  width: 100%;
-  position: relative;
-}
-
-#the-map {
-  height: 100%;
-  width: 100%;
-}
